requirements.txt
Flask~=3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
Flask-Sock
SQLAlchemy==2.0.37
flask-mail~=0.10.0
flask-migrate~=4.1.0
pymysql~=1.1.1
python-dotenv~=1.1.0
genai
google-genai
tiktoken --only-binary=:all:
fpdf2~=2.8.2
apscheduler~=3.11.0
pi-ina219
RPi.GPIO
websocket-client
flask-limiter

Werkzeug~=3.1.3